version 1.0 done

+++ sort_dwnlds.py
@@ -2,24 +2,27 @@ import os
 import shutil
 
 user = os.getenv("USER")
-dir = "/Users/"+user+"/Desktop/hat/"
-# dir = "/Users/"+user+"/Downloads/"
+# dir = "/Users/"+user+"/Desktop/hat/"
+dir = "/Users/"+user+"/Downloads/"
 
 try:
     os.chdir(dir)
     for files in os.listdir(dir):
         if not files.startswith('.'):    # Ignoring hidden files
-            print files
             if os.path.isfile(files):
-                extn = files.split(".")[1]
-                print extn
+                extn = os.path.splitext(files)[1].replace('.', '')
                 dirname = extn.upper()
                 if os.path.isdir(dir+dirname):
-                    shutil.move(dir+files, dir+dirname)
+                    if os.path.isfile(dir+dirname+"/"+files):
+                        print "Could not move "+dir+files+" it already exist in "+dir+dirname+"/"
+                        continue
+                    else:
+                        shutil.move(dir+files, dir+dirname)
 
                 else:
                     os.mkdir(dir+dirname)
                     shutil.move(dir+files, dir+dirname)
+    print "All Done!"
 
 except IOError:
     print "Error in reading file"
